Migrate CLI commands to `llm_router_cli` package and remove old `llm_router_api` CLI implementations.
