--- llm_router_api/cli/__main__.py
@@ -1,7 +1,0 @@
-"""Allow ``python -m llm_router_api.cli``."""
-
-import sys
-
-from llm_router_api.cli import main
-
-sys.exit(main())
