--- llm_router_api/cli/commands/__init__.py
@@ -1,7 +1,0 @@
-"""
-CLI command module for llm-router.
-"""
-
-from .auth import register_auth_subparser, main as main_auth
-
-__all__ = ["register_auth_subparser", "main_auth"]
