+++ llm_router_cli/cli/__main__.py
@@ -1,0 +1,6 @@
+"""Allow ``python -m llm_router_cli.cli``."""
+
+import sys
+from llm_router_cli.cli import main
+
+sys.exit(main())
